Make transposition node smaller
Dramatically reudce memory usage of the transposition table, trying to
improve cache hits. It's still huge, but this helps a little bit.

+++ search.h
@@ -3,7 +3,7 @@
 
 #include "types.h"
 
-void search_force_max_depth(const int depth);
+void search_force_max_depth(const int8_t depth);
 Move search_find_move(Bitboard *board);
 
 #endif
